add performance tests and refactor all other tests
